+++ java/client/src/org/openqa/selenium/remote/CapabilityType.java
@@ -39,6 +39,7 @@ public interface CapabilityType {
   String HAS_NATIVE_EVENTS = "nativeEvents";
   String UNEXPECTED_ALERT_BEHAVIOUR = "unexpectedAlertBehaviour";
   String ELEMENT_SCROLL_BEHAVIOR = "elementScrollBehavior";
+  String HAS_TOUCHSCREEN = "hasTouchScreen";
 
   String LOGGING_PREFS = "loggingPrefs";